Fix Static Analysis Win64 (MSVC) - Arithmetic overflow
#jira UE-156593
#rb richard.malo
#preflight none
#rnx

[CL 20652140 by JeanFrancois Dube in ue5-main branch]

+++ Engine/Source/Runtime/Engine/Private/WorldPartition/WorldPartitionRuntimeSpatialHash.cpp
@@ -897,7 +897,7 @@ void UWorldPartitionRuntimeSpatialHash::DumpStateLog(FHierarchicalLogArchive& Ar
 						LevelActorCount += Cell->GetActorCount();
 					}
 				}
-				LevelsStats.Add({ LevelCellCount, (StreamingGrid.CellSize << Level), LevelActorCount });
+				LevelsStats.Add({ LevelCellCount, ((int64)StreamingGrid.CellSize << (int64)Level), LevelActorCount });
 				TotalActorCount += LevelActorCount;
 				++Level;
 			}
